Added some stubs for getting information about an authentication plugin, although they may be removed.
Check for whether the system is actually installed instead of checking for
if the version is available, which allows the install to proceed without
the user being bothered by pesky details such as having to log in when no
accounts are available.

Get the auth script from the new location, which will change to
safe_require when I get hold of it.