Broke out the pagination building into the view class, so the artefactchooser json script can also use it.
There is a bit of a hack to make sure the javascript for the paginator is available, by storing it in a global variable and accessing that later in the appropriate function. This will probably cause problems if more than one artefact chooser is being rendered on the same page, but I think that support for that should be disabled anyway. It will be fixed later when Pieforms goes to a more OO element format.