Two bugfixes, one small, one large.
The page was being first hit using the 'id' parameter, pieforms was building the form based on this. But the hidden element in the form was called 'viewid'. When the page was being hit again after form submission, the form was built again assuming 'viewid' held the view ID, which it did not anymore. Thus the form was built with and ID of 0, and the submit handler behaved appropriately - by creating a new view.

The small bug was that the 'save' button said 'new' when you were saving an existing view.