Changes to the way we handle config.php and the debconf questions There are two files in /etc/mahara: config.debconf.php and config.php.  Both of these have symlinks in /var/www/mahara.
config.debconf.php is managed by debconf and is overwritten everytime the package is reconfigured.  It is not marked as a config file.

config.php includes config.debconf.php first in order to allow users to overwrite the debconf settings.  It is marked as a config file so that it's not overwritten automatically at configure time.