Prevent a fatal error caused by adding the entireresume blocktype then visiting your resume for the first time.
The 'contactinformation' blocktype is attempted to be added to the entireresume blockinstance that has been created, causing badness as the blocktype hasn't actually been committed yet. Move the part of the commit where it's added to the database before the part where it's added to any views.

Also make sure that contact information isn't shown in a view as just a <table></table> if it has no content.