Added the ability for the paginator to fire events when it changes things.
Given that the paginator object is created when the pagination is put onto the page, and destroyed when the pagination changes (i.e. when the user changes a page), a PaginationProxy class is used to collect observers, and the pagination class sends events to it.

Currently the only event fired is when the page changes. This is used by the artefact chooser when working out what checkboxes are selected on previous pages.