Added a helper object to work out what checkboxes have been chosen on each page of the artefact chooser. Fixes #1718.
The fix involves using a helper object to remember what has been selected and what has been seen on each page. The helper munges hidden inputs around so that the correct data is sent back to the server, where 'correct' means the list of checkboxes checked on all pages that were visited, and the list of all checkboxes that were on visited pages.

It's quite a lot of javascript to send back with each request, so could be broken out somehow into a separate file at a later date.