Fix problem where only the first reply to any post was being shown for me. Now all posts in a thread are shown.
The symptoms I saw only happened because I replied in succession to the last post.

The problem was caused by looping over a reference variable that was then passed to a recursive function. I think internally PHP moved the array pointer for it to the very end when it reached the first time there were no children, then when it was coming back up through the recursion all the loops were being finished.