Applied the same feedback improvements to viewing an artefact that were previously applied to viewing a view.
Also make it impossible to submit a file as an attachment when viewing an artefact. Digging around on production suggests this feature isn't used, and it's buggy because if the user chooses to upload a file, then the feedback is placed on the view rather than the artefact they were viewing.