Make each 'add friend' button on user search pass through the query if it's there.
This is so that the same query will be run, and we can guarantee that the pieform will be available so processing can occur.

This is a bit of a hack. It would be better to do the user search results in one big form, and toggle on the submit button pressed to work out what to do.
(cherry picked from commit edba66f3f441587a805bdd23781e533fc93ea3e8)