Use correct validation on institution name when creating via promiscuous mode. Fixes #2092.
Previously, the code was only making sure whitespace was removed, but other parts of Mahara depend on this field having only alphabetical characters.

Now the code makes sure only alphabetical characters are in the string, and will cycle through the letters a - z attempting to generate a new institution name if the one generated is already taken.

If someone needs more than 26 peers all with the same name... well tough :)