Make most blocktype titles editable.
This patch makes most blocktypes have no title set by default when created. For those that used to set their own titles, if no title is set after the configuration form is submitted, then they get the chance to set the title.

So for example, the RSS block will start with an empty title, but if the user doesn't set it, the title will be the title of the feed that is being syndicated.

Blocks like 'textbox' still keep their defaults, because they don't have any special case way of setting their own titles.