Improved handling of parent/child authentication so duplicate users are not created.
This comes from a report on the forums: http://www.mahara.org/node/155

The problem stemmed from an apparent misunderstanding in the code about the purpose of the 'authinstance' column in the usr table. It is for the authentication instance the user uses to log in, and is normally the parent if they are using a child/parent authentication method. However, if the authentication method doesn't have a parent, it will be the child.

This means we need to change checks in a couple of places to look for the parent _or_ child, rather than just one of them. Otherwise, some parts of the code believe that users don't exist when in fact they do, etc.