Allow users to specify paths on the filesystem to search for language packs.
If a user has checked out a language pack repository, they can add the path of it to a new configuration variable, 'langpacksearchpaths', which is an array of paths to search.

Helpfiles are not handled yet.

This can be tested by checking out the en_AR.utf8 language pack and adding it to langpacksearchpaths. You should get an option to use language 'Avast!'.