Allow profileicon thumbnails to be cached, at least for 10 minutes.
We can't cache them for too long, because the user could change them at any time. But at least a short time is better than none, because they can be requested in batches quite a lot, for example when paging through search results.

On the profile page for the user, use 'profileiconbyid', which means that the correct image will always be shown, even when the user changes it.