Implement a cron job to delete old session files. Fixes #570.
We haven't been clearing out session files since I first chose to make us hash the session directory back in 2006. Talk about a timebomb...

The cron job uses `find' and `xargs' to do the removing. These tools are required on debian (as part of findutils), and are installed in /usr/bin. I haven't bothered with a configuration directive for specifying a path to them for now, but that might be necessary later.