Add an option - usersuniquebyusername - that allows users from multiple institutions to SSO in to one Mahara account. Tentative fix for #2243.
If this option is turned on (only in config.php), then when users SSO in from any XMLRPC authentication instance, they will be considered the same user if they have the same username. This differs from current behaviour, in that if the users' authinstances are different then they are considered different users.

If this option is turned on, no institution in the system is allowed to have self registration turned on. Also, users must be able to belong to multiple institutions.

This option added thanks to Howard Miller/Glasgow University.