(#3063) part 2: Hide blocktypes that don't make sense when making a View.
This means that blog/profile/resume blocktypes don't show up in
group/institution/site Views.

This is controlled through a new Blocktype method - allowed_in_view.
Various code had to be changed to listen to this method, mostly around
the display of the category browser.

When making a template, all blocktypes are available.

Part 3 is to change the configuration forms to be more sensible when
editing a template, and then testing what happens when blocktypes are
rendered somewhere that they "shouldn't" be - e.g. a My Friends block in
a group View.