Remove a noisy log line.
It's noisy because the edit profile page wants lots of helpfiles that haven't been written, and it's too close to a release to set about writing them.

In practice it would be more useful along with other debug lines that tell a translator that they need to translate certain help files.