Add a CSS file that is only hooked in if javascript is enabled.
We can use this to apply certain styles when javascript is around - most notably, to have stuff that is shown when javascript is off, but hidden by default when it is on.

There are two classes provided to help with this: js-hidden and nojs-hidden. They hide the element they're applied to when javascript is on or off respectively.