Unify content type handling in LEAP export, and switch to 'html' content type.
We're not exporting as xhtml any more - that work would unfortunately conflict with future work we have planned.

Now there's a new method, get_content_type, which elements should define to say what type they should be. The default implementation of get_content is now applicable in many more cases. All escaping and cleaning is done in the template, where it should be.

The system still supports exporting content with type=xhtml - we're just not going to use it.