Look for static files in artefact plugins, and do it by theme as well.
This required splitting up the export polling of artefact plugins into two steps - one where we establish what stylesheets are going to be needed, and the second for actually doing the export. This is because the second step writes out templates, so we have to know the stylesheets before this.

This required re-calculating the progress bar, those changes are reflected here too.