[UPSTREAM] Prevent internal notifications being 'stolen' from users in rare circumstances.
This was happening because a variable ($userdata->internalid) was being used
for two purposes - one to hold the ID of the notification as put into the
notification table, and also to see if the notification needed to be updated
after having been inserted for any reason (either the notification was sent by
another method successfully so it needed to be marked as read, or the URL set
in the table needed to change post-insert).

The particular method this bug is in is called repeatedly in a loop, once for
each user who is getting a notification. Unfortunately, $userdata->internalid
comes from converting the object to a stdclass, and the previous internalid is
set on the class too. So if the notification didn't need to be marked as read
or have its url changed (e.g. it's a forum post notification going straight to
the activity log), then it would still be updated - this would have been
harmless, but the update would ovewrite the user who owned the notification.

Will make a few more patches to make the code more consise and limit the chance
of more damage. This problem took me hours to debug...