Don't link up images that are already linked up (haxy).
Per report in the forums from Michael Schwab:
http://mahara.org/interaction/forum/topic.php?id=893

This won't work if there's tags between the <a> and the <img> - but possibly
the regex can be tweaked to allow that. In my testing, it handles his case at
least, which should hopefully be good enough for 95% of image linking usage out
there.