Calculate display order for imported resume items.
We do this by looking for whether the items are connected to one of the
special Mahara groupings. If they are, we read the display order and use
it - otherwise, the display order isn't used, as we assume the display
order is for some other selection that is irrelevant to Mahara. This
assumption might need to change in future.

Signed-off-by: Nigel McNie <nigel@catalyst.net.nz>