Very, very basic View import.
Imports all content into one big text box.

Manages to import title, summary, ctime and mtime too. Still have
ownerformat to do.

The next step after this is to try and detect if we're importing a
Mahara view, and if so, create the right blocks and artefact links for
it.

Signed-off-by: Nigel McNie <nigel@catalyst.net.nz>