Import UI: Allow user accounts to be created by uploading a LEAP2A file.
Still some minor things to clean up, but it works reasonably well.

I experimented with a horizontal UI.. I think it would be better if a
graphic designer could make it more obvious that it's a 1-2-3 process.
But everyone I've tried it on has managed to create a user account.

Still todo:
 - progress indication for when a leap2a file is uploaded
 - some minor tweaks to the way some data is imported
 - needs documentation on the wiki and a help icon or two

Signed-off-by: Nigel McNie <nigel@catalyst.net.nz>