Initial version for the HTTPS login patch
There is a new optional setting called $cfg->httpswwwroot. The reason we
have a new setting instead of simply substituting 'http://' with
'https://' in $cfg->wwwroot, like Moodle currently does, is that this
prevents people from using non-standard ports in their HTTPS setups.
This may be necessary if you want to run both Moodle and Mahara on the
same host/ip and use HTTPS logins for both of them.

Signed-off-by: Iñaki Arenaza <iarenaza@eps.mondragon.edu>