Make LeapExportElement::assign_smarty_vars throw an exception if called without the artefact field being set.
It seems this class can be used in two ways. In one way, it represents
one artefact being converted to an entry. The other way is where
something that isn't an artefact is being converted (e.g. a resume
"psuedo artefact"). Which is a bit naff, but no time to fix right now.
Instead, make this method die horribly if it's called directly. All
current callers override this method without calling the parent anyway.

Signed-off-by: Nigel McNie <nigel@catalyst.net.nz>