Ensure all config values for LEAP2A export/import are json_encoded in wrapper arrays.
It turns out I was relying on a nasty behaviour quirk of json_decode in PHP,
namely that it'll just pass strings through - in some cases. This is horribly
out of touch with the JSON specification, see these PHP bugs:

http://bugs.php.net/bug.php?id=38680 (where they broke it)
http://bugs.php.net/bug.php?id=46518 (people complaining about it)

The solution is to wrap all configuration fields when exporting as LEAP in an
array and then json_encode that. json_decode will handle that properly in all
versions of PHP.

Signed-off-by: Nigel McNie <nigel@catalyst.net.nz>