Allow all types of resume artefact in the resumefield blocktype to be imported.
This was broken because contactinformation is a "fake" artefact that
just ties together some profile artefacts, so we have to create one at
import time for the user. We also have to export the fields by name
instead of artefact ID in the Mahara config, as the contactinformation
artefact isn't exported at all.

Signed-off-by: Nigel McNie <nigel@catalyst.net.nz>