Prioritisation of per-institution themes
If a user is a member of multiple institutions, they would have typically
received the theme for the last institution (as returned in an unsorted
order from lib/user.php->load_user_institutions).

This patch introduces a check to determine the user's authentication
instance and the institution associated with that instance.
If a theme is set for that institution, this is used in priority.

Signed-off-by: Andrew Robert Nicols <andrew.nicols@luns.net.uk>