Add plugin specific javascript capability.
Blocktype plugins can now specify javascript files
to be called at the top of the view. The files should
be specified relative to lib.php, and must be instance
independent. Each file will be called once per blocktype
in a view, not once per block instance.

Blocktypes can also specify javascript to be loaded when
instances of the blocktype are being configured.