artefact/file/image: maintain image type when resizing (bug #495234)
The image resizing code used to convert all images to PNG. This would
cause problems in certain cases when browsers only looked at the extension
to figure out the file type.

For example, exporting a view containing a resized JPG file would create a
PNG file with a .jpg extension. Using IE on Windows to display the exported
page would file as the mime/type was inferred from the extension.

Signed-off-by: Francois Marier <francois@catalyst.net.nz>