Consider quota when determining the 'maxuploadsize' for My Files.
If a user has a quota which is lower than post_max_size or
upload_max_filesize, then this should be shown as the maximum upload size.

The actual checking of this is already handled by the upload function for
the pieform.
(cherry picked from commit af0facaa676e716a29163c79e112fda0e54817fb)

Signed-off-by: Evan Goldenberg <evang@catalyst.net.nz>