Leap importer: fixed the default fallback import strategy a bit
Since the blog is the default fallback importer, make it be aware of
selections and import them as blogs, rather than blogposts.

It's still necessary to have a strategy for importing these entries as
blogposts as well, in case some of the child entries have a higher
weighted strategy as the fallback, because in that case, they'll be
imported using that strategy and the selection would be lost.

Signed-off-by: Penny Leach <penny@mjollnir.org>