usability: always show list of blogs, even if there is only one
Also, remove the "Add Blog" button from the view blog page, where it is
out of context. This simplifies the navigation of the blogging features
for user. Even when they only have one blog, being presented with a list
makes the ability to have more than one blog immediately apparent.

Signed-off-by: Evan Goldenberg <evang@catalyst.net.nz>
Signed-off-by: Francois Marier <francois@catalyst.net.nz>