Remove the connection between a user's profile and profile view
The profile view is just a special view, and really has no obvious
connection to the user's profile information. Linking the two blurs the
central distinction between content and views in Mahara.

This patch loosens that connection by:
    - Removing links from the edit profile page to the profile view
    - Removing links from the profile view to the edit profile page
    - Showing the profile view in the My Views list

While the profile view is now treated like other views, it cannot be
deleted, and its access cannot be changed. Links to these features do
not appear for the profile view in the My Views list.

Signed-off-by: Evan Goldenberg <evang@catalyst.net.nz>