Add view submission to group
This gives users a list of all views they haven't yet submitted to any
group if they're in the group/view.php page for a group which is of a
submittableto group type, and to which they haven't already submitted a
view to.
I've added a duplicate language string for 'Submit' as I suspect the submit
button text won't work for translations in this context.

Signed-off-by: Andrew Robert Nicols <andrew.nicols@luns.net.uk>
Signed-off-by: Evan Goldenberg <evang@catalyst.net.nz>