Refactor view/index.php to correctly set view options for each view (Fixes #627966).
I've also removed some old variables which don't seem to be used any more.

To fix this bug, I've moved the loop through of the users views. Whereas
before, the loop was only entered into if there were course groups, the
loop is always hit now.

While this is less efficient if the site has disabled public profile views,
and the user has no course groups, it's more effective if either of these
cases is true, and allows for easier expansion in the future.

Signed-off-by: Andrew Robert Nicols <andrew.nicols@luns.net.uk>