uploadmanager: lowercase file extensions to fix flowplayer bugs
We keep track of the original file extension of the file to work
around some Internet Explorer bugs. However, given that's for
Windows, we have no need to preserve the orignal case of the
file extension.

Converting it to lowercase at upload time will prevent a whole
class of bugs with poorly-tested third-party software.

Bug #675047

Signed-off-by: Francois Marier <francois@catalyst.net.nz>