artefact/comment: fix deletion of feedback on views and artefacts
This fixes a bug which cause any delete button next to feedback
artefacts to delete the first feedback artefact on that page
instead of the one that was selected.

Since the submit button was missing a name, the view never entered
the codepath where the right comment ID was given as an argument
to the delete form. Instead the delete action was handled (with
the wrong comment ID) during the setup of the pieform.

Signed-off-by: Francois Marier <francois@catalyst.net.nz>