Automatically generate block titles when they are rendered
The blocktype's get_instance_title function was previously called when
blocks were saved.  By moving the call to happen when the block is rendered,
it's possible to (for example) automatically update the title of a blogpost
block by changing the post title.

This change will have no effect on existing blocks, because
get_instance_title is only called when the title stored in the db is empty.

Change-Id: I599108e7b96c80475600d21c3b2054096ce162c0
Signed-off-by: Richard Mansfield <richard.mansfield@catalyst.net.nz>