Hide title in block config form for autogenerated block titles
When block titles are empty and the block can generate its own title,
hide the title field in the configure form and add a link to expand it.
Don't rely on block title being != blocktype title to determine whether
the block title has been edited, because changing languages breaks this.

Change-Id: I173ff76176461db1ce24ad8e46d84a1783cb6dee
Signed-off-by: Richard Mansfield <richard.mansfield@catalyst.net.nz>