Don't put other users' files in html export of files area (bug #688304)
Html export creates a browseable version of the exporting user's files
area.  Sometimes files owned by other users or groups are included in
the export, and the html export attempts to include those in the file
listing.  The problem is particularly bad when there are name conflicts,
for example when the user includes a group folder in a view, and that
folder has the same name as one of the user's personal folders.

Change-Id: I4bbdf3127735cb9c4af13055546d2b08866d47cf
Signed-off-by: Richard Mansfield <richard.mansfield@catalyst.net.nz>