Html export: add files not owned by exporting user to special directory (bug #688304)
Sometimes an html export is allowed to include files that are not part of the
exporting user's portfolio, e.g. when the user has permission to publish a
group file in a personal view.  Make sure these files are included in the
export tarball, but keep them separate from the user's personal files.

Change-Id: Ic9bb7a5d059c9c39d0fb71e58e843dc048cbdfb8
Signed-off-by: Richard Mansfield <richard.mansfield@catalyst.net.nz>