Move display_name html-escaping out of templates and into the dwoo function
There is an existing display_name dwoo function which calls the
user/lib.php function with the same name.  This change wraps that
call in hsc(), and removes explicit escaping from all the templates
that use it.  In future, template authors won't have to remember to
put "|escape" after every call to display_name.

Change-Id: Ia8a614f6679f6cfa1b17508b2c820717cc311ef4
Signed-off-by: Richard Mansfield <richard.mansfield@catalyst.net.nz>