googleapps block: use https images when needed
On Mahara sites served over HTTPS, the Google Books, Calendar,
Docs and Maps logos were seen by the browser as insecure content.

This patch uses the HTTPS version of the logos when needed.

It also always links to the secure versions of Docs and Calendar
just to ensure that our users are directed to the secure versions
of these Google services.

Change-Id: I40aefc4798cde90bd31c4ac711cbb9d2a7b1b602
Signed-off-by: Francois Marier <francois@catalyst.net.nz>