Feature: Generate a sitemap (bug #793815)
Code to generate a sitemap generation code, using the sitemaps.org
standard/protocol. The default generation is daily, with a
full generation done on the first of the month. The full
generation can also be forced by passing 'true' to the constructor.

Sitemap file(s) are placed in a 'sitemaps' directory just off
the docroot, with an index file placed in the docroot itself. If
the 'sitemaps' directory does not exist, one will be created.

More information can be found on
https://wiki.mahara.org/index.php/Developer_Area/Specifications_in_Development/Sitemaps

Change-Id: Ib28983ca1c3b5ff06a6e603f402460531b987067
Signed-off-by: Darryl Hamilton <darrylh@catalyst.net.nz>