Add usr table indexes for faster searches
Search is much faster in Postgres with these indexes on the lowercase
profile fields.

In MySQL we can't create the indexes without changing the types of the
column from text to varchar.  An upgrade to do this would have to be
destructive, and would require admin notification.  It's not worth doing
that, because MySQL cannot create indexes on functions of columns,
and we could not create the lowercase indexes that would be useful for
user searches.

Change-Id: I9ffb6ce02a5f3e8fc2cf158424b37b754edd8d34
Signed-off-by: Richard Mansfield <richard.mansfield@catalyst.net.nz>