Reset the site language immediately after config is loaded (bug #787123)
Because the $lang variable is declared as static in current_language(), a
call to get_string in the sanity checks can set the language for the
entire request, even before the site language has been read in from the
config table.  This ensures the language is reset immediately after config
has been read.

Change-Id: Ic81dfaedd70de75d93a6d38dd89915eeabcccd6e
Signed-off-by: Richard Mansfield <richard.mansfield@catalyst.net.nz>