Revert "Remove wwwroot from the database (bug #780177)"
This reverts commit fb38dd955ceed3601da954dd5b2e55b682e1ac03.

It turns out that we do need the wwwroot to be in the database
because the code that sends forum posts via email runs in php
cli mode and is needs to get the hostname of the site to embed
in the headers.

While not specifically mentioned in bug #794490, it was spitting
out warnings in the logs.

Conflicts:

	htdocs/lib/db/upgrade.php
	htdocs/lib/version.php

Change-Id: Icb734f939889453e80b53647748976648ee18419
Signed-off-by: Francois Marier <francois@catalyst.net.nz>