Allow updates to existing users by csv upload
Adds an 'Update Users' checkbox to the CSV upload page.  When checked,
users on the site whose usernames appear in the CSV file will have their
details overwritten by the values in the CSV file.

To be updated, existing users must be in the same institution as the one
specified in the upload form.

Change-Id: Ia47f7034adbade21dd26f455c356f58ed7bbf730
Signed-off-by: Richard Mansfield <richard.mansfield@catalyst.net.nz>