Make index_exists more useful when called on indexes with no columns (bug #796102)
ddl's index_exists() function tries to check whether an index exists on a table
by comparing the column names in the index passed to it with all the indexes on
the table.

Indexes on functions of columns all appear as having no columns, so the behaviour
is not very useful.  If a table has any index on a function, index_exists will
always return true when passed an index with no columns, whether or not its name
matches any existing index on the table.

This change modifies find_index_name() so that it treats indexes without columns
as a special case, and matches on their names rather than on the list of columns.

Change-Id: Ic0d664dff77fe5a220a9f9563975df1c87b18e06
Signed-off-by: Richard Mansfield <richard.mansfield@catalyst.net.nz>