Fix bug with pagination on journals page
Bug #800575

When pagination was used on the journals page, the user
could not select any page other than the current. An error
log occured if they did try.

This fix adds a requires clause to the json script that is used
by the pagination. Fixing this also brings up the issue in bug #795425

Change-Id: I17516109f450e2044e4276e3f1a83dec2488bd26
Signed-off-by: Hugh Davenport <hugh@catalyst.net.nz>