Remove requested/invited users from the admin search
See bug #799617

This change fixes the pagination on the admin user search page for
institutional admins with more than one institution when users can
belong to multiple institutions.

As part of the fix, the behaviour of the page has been changed so
that requested and invited users are no longer displayed.

Institutional admins are already able to find users with open requests
on the institution members page, and will soon be able to find invited
users on that page too.

Change-Id: I2460bcf54e0e0ca96a51dd592c04aa0a2229cc7d
Signed-off-by: Richard Mansfield <richard.mansfield@catalyst.net.nz>