file browser: remove confusing UI in folder block (bug #808593)
The broken download link has been removed when a directory is
selected (both in the AJAX call and the form setup).

This also fixes the display of the "No files found" message to
make sure that it's only displayed when the selection list is
empty and to hide it in the case of folder selection.

Change-Id: I1a2b2e5d6ea7cf803b7c9d23e9a91186a39d9b03
Signed-off-by: Francois Marier <francois@catalyst.net.nz>