Save textbox content to the artefact table (bug #736665)
When a text box block is saved, a new html artefact is created,
or an existing html artefact updated.  The 'Block content'
field from the form is saved into the description field of the
artefact, and the artefact title is overwritten with the block
title.

This patch includes an upgrade to convert the contents of all
existing textboxes on the site to html artefacts.

Change-Id: I321735de5c0e4dc5768222f02c8e7e1bc716c237
Signed-off-by: Richard Mansfield <richard.mansfield@catalyst.net.nz>