file_mime_type: increase use of hardcoded magic file (bug #809216)
Mahara's mimetype detection function wasn't working properly and
so it would not be able to set the right mimetype when a browser
(typically IE) sets it to a generic value.

The workaround in 821a693c8da162b91b0715eaee771b09d19de601 is not
sufficient because symlinks, like directories, don't work but the
is_file() function will resolve them and may return true.

  http://nz.php.net/manual/en/function.is-file.php#50742

Change-Id: Ie8485ee10e5142a5f00032116c5492b9778dae51
Signed-off-by: Francois Marier <francois@catalyst.net.nz>