Remove profile icon from session on delete (see bug #805742)
Commit 0c9774b9 added a foreign key constraint on the profile icon
field of the user table.  But when a profile icon is deleted, this
field isn't updated in the global $USER object, unless the icon is
deleted from the profile icons page, so the bad profileicon can hang
around in the session, and then cause an sql error the next time the
usr table is updated from $USER.

This change will fix that problem when a user deletes their own
profile icon, however it may still be possible for the error to occur
if for example the user is logged in twice.

Change-Id: I2f8792fcd392be351b8ee9ac77337fd41b046a3d
Signed-off-by: Richard Mansfield <richard.mansfield@catalyst.net.nz>