Allow paginator.js to paginate contents of non-table elements
Paginator.js expects replacement html returned from an ajax call to
be a string containing only table rows, which it then inserts into
an existing table on the page.

Recent updates to the myfriends block templates replaced the tables
with divs, breaking the pagination of the block.

This change modifies the paginator so that it can cope with divs as
well as tables.

Change-Id: Ie2d6139e3e57cab3d5900f47892f6b1745106edc
Signed-off-by: Richard Mansfield <richard.mansfield@catalyst.net.nz>