Generate new parent folder for each new view copied (bug #820180)
Files copied during view copying were not dealt with correctly
when more than one view was being copied in a single request.
The copying process automatically creates a folder in which to
put any files it needs to copy, but the id of this folder was
declared as static, so it didn't work once copying of multiple
views is allowed.

There is a similar problem for blogs created during view copies
(to hold copied blogposts).  I'm not sure if there's a way to
copy multiple views containing blogposts that need copying (it
can't happen in institution views).  But the default blog creation
function is updated in this patch anyway.

Change-Id: Id8ea2504772f49ce8f301cfc19d41ba20b587e9d
Signed-off-by: Richard Mansfield <richard.mansfield@catalyst.net.nz>