Add an artefactchooser to the textbox config form (bug #736665)
The configuration form's tinymce field allows the user to edit
the currently selected html artefact.  A new link below the content
field displays an artefactchooser which, allows the user to select
from their existing html artefacts.  On selecting an artefact, the
tinymce contents are replaced with the contents of the selected
artefact.  Also displays a message to the user whenever there are
other blocks containing the same artefact as the one they've
selected.

Change-Id: I2ad655714429252ed3877e6589c37c6bbcaad37e
Signed-off-by: Richard Mansfield <richard.mansfield@catalyst.net.nz>